For load-config :carica/json, directly open an io/reader on resource
.openStream is a java.net.URL method, which means, if you pass "something io/reader could
handle, but not a URL" you get a "No matching field found: openStream" exception.

For example, a java.net.URI, which io/reader handles fine, threw that exception. Basically,
the assumption was that load-config was only ever called on a resource directly returned
from io/resource (which returns java.net.URL because that is what ClassLoader/getResource
returns), but that's not necessarily always true for anyone calling load-config with json
from a java.net.URI, or a java.io.File, or even a string -- "(carica/load-config
"/path/to/some.json")" would also throw.